Move site-header to components folder
