+++ src/page-multi-facility/FacilityPage.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from "react";
 import styled from "styled-components";
 
+import SiteHeader from "../components/site-header/SiteHeader";
 import Loading from "../design-system/Loading";
 import { EpidemicModelProvider } from "../impact-dashboard/EpidemicModelContext";
 import { useLocaleDataState } from "../locale-data-context";
 import useScenario from "../scenario-context/useScenario";
-import SiteHeader from "../site-header/SiteHeader";
 import { FacilityContext } from "./FacilityContext";
 import FacilityInputForm from "./FacilityInputForm";
