+++ src/page-multi-facility/MultiFacilityPage.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 
+import SiteHeader from "../components/site-header/SiteHeader";
 import Loading from "../design-system/Loading";
 import { useLocaleDataState } from "../locale-data-context";
 import useScenario from "../scenario-context/useScenario";
-import SiteHeader from "../site-header/SiteHeader";
 import CreateBaselineScenarioPage from "./CreateBaselineScenarioPage";
 import MultiFacilityImpactDashboard from "./MultiFacilityImpactDashboard";
